+++ src/Writer.php
@@ -82,7 +82,8 @@ class Writer
     {
         return $this->config->queriesEnabled() &&
             $query->time() >= $this->config->queriesMinExecTime() &&
-            preg_match($this->config->queriesPattern(), $query->raw());
+            preg_match($this->config->queriesIncludePattern(), $query->raw()) &&
+            !preg_match($this->config->queriesExcludePattern(), $query->raw());
     }
 
     /**