+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # CHANGELOG
 
-## [v0.9.1 (Unreleased)](https://github.com/onlime/laravel-sql-reporter/compare/v0.9...main)
+## [v0.9.2 (Unreleased)](https://github.com/onlime/laravel-sql-reporter/compare/v0.9.1...main)
 
+## [v0.9.1 (2021-06-03)](https://github.com/onlime/laravel-sql-reporter/releases/tag/v0.9.1)
 
-## [v0.9 (2021-06-01)](https://github.com/onlime/laravel-sql-reporter/releases/tag/v0.9)
+### Added
+- Added new config param `queries.exclude_pattern` (env var `SQL_REPORTER_QUERIES_EXCLUDE_PATTERN`) to narrow down queries to be logged without bloating include pattern regex.
+
+### Changed
+- Renamed `SQL_REPORTER_QUERIES_PATTERN` env var to `SQL_REPORTER_QUERIES_INCLUDE_PATTERN`
+
+## [v0.9 (2021-06-02)](https://github.com/onlime/laravel-sql-reporter/releases/tag/v0.9)
 
 - Initial release